Create Motion Detecting Vehicle
The goal of our design was to build a car that could detect objects in front of it and maneuver its path accordingly. We decided to use a photocell combined with a RedBoard, Arduino code and a car chassis to accomplish this task. We used an H-Bridge to drive and control the motors.